+++ redux-anecdotes/src/App.js
@@ -1,29 +1,12 @@
 import React from 'react'
-import { useSelector, useDispatch } from 'react-redux'
-import { createAnecdote, voteForAnecdote } from './reducers/anecdoteReducer'
 import AnecdoteForm from './components/AnecdoteForm'
+import AnecdoteList from './components/AnecdoteList'
 
 const App = () => {
-  const anecdotes = useSelector((state) => state)
-  anecdotes.sort((a, b) => b.votes - a.votes)
-  const dispatch = useDispatch()
-
-  const vote = (id) => {
-    dispatch(voteForAnecdote(id))
-  }
-
   return (
     <div>
       <h2>Anecdotes</h2>
-      {anecdotes.map((anecdote) => (
-        <div key={anecdote.id}>
-          <div>{anecdote.content}</div>
-          <div>
-            has {anecdote.votes}
-            <button onClick={() => vote(anecdote.id)}>vote</button>
-          </div>
-        </div>
-      ))}
+      <AnecdoteList />
       <AnecdoteForm />
     </div>
   )
