+++ redux-anecdotes/src/components/AnecdoteList.js
@@ -1,0 +1,35 @@
+import React from 'react'
+import { useDispatch, useSelector } from 'react-redux'
+import { voteForAnecdote } from '../reducers/anecdoteReducer'
+
+const Anecdote = ({ anecdote, handleClick }) => {
+  return (
+    <div key={anecdote.id}>
+      <div>{anecdote.content}</div>
+      <div>
+        has {anecdote.votes}
+        <button onClick={handleClick}>vote</button>
+      </div>
+    </div>
+  )
+}
+
+const AnecdoteList = () => {
+  const dispatch = useDispatch()
+  const anecdotes = useSelector((state) => state)
+  anecdotes.sort((a, b) => b.votes - a.votes)
+
+  return (
+    <div>
+      {anecdotes.map((anecdote) => (
+        <Anecdote
+          key={anecdote.id}
+          anecdote={anecdote}
+          handleClick={() => dispatch(voteForAnecdote(anecdote.id))}
+        />
+      ))}
+    </div>
+  )
+}
+
+export default AnecdoteList
